+++ extensions/reclaim-ai/src/hooks/useEvent.tsx
@@ -7,7 +7,7 @@ import { NativePreferences } from "../types/preferences";
 import { axiosPromiseData } from "../utils/axiosPromise";
 import { formatDisplayEventHours, formatDisplayHours } from "../utils/dates";
 import { filterMultipleOutDuplicateEvents } from "../utils/events";
-import { parseEmojiField } from "../utils/string";
+import { stripPlannerEmojis } from "../utils/string";
 import reclaimApi from "./useApi";
 import { ApiResponseEvents, EventActions } from "./useEvent.types";
 import { useTask } from "./useTask";
@@ -42,7 +42,7 @@ export const useEvents = ({ start, end }: { start: Date; end: Date }) => {
     }
   );
 
-  if (error) throw error;
+  // if (error) throw error;
 
   return {
     events: filterMultipleOutDuplicateEvents(events),
@@ -69,15 +69,18 @@ export const useEventActions = () => {
             hoursFormat: meridianFormat,
           });
 
+          // console.log('=> event.sourceDetails?.title', event.sourceDetails?.title)
+          // console.log('=> event.title', event.title)
+
       const realEventTitle = event.sourceDetails?.title || event.title;
-      return `${hours}  ${parseEmojiField(realEventTitle).textWithoutEmoji}`;
+      return `${hours}  ${stripPlannerEmojis(realEventTitle).textWithoutEmoji}`;
     },
     [currentUser]
   );
 
   const handleStartHabit = async (id: string, title: string) => {
     try {
-      await showHUD("Started Habit: " + parseEmojiField(title).textWithoutEmoji);
+      await showHUD("Started Habit: " + stripPlannerEmojis(title).textWithoutEmoji);
       const [habit, error] = await axiosPromiseData(fetcher(`/planner/start/habit/${id}`, { method: "POST" }));
       if (!habit || error) throw error;
       return habit;
@@ -89,7 +92,7 @@ export const useEventActions = () => {
 
   const handleRestartHabit = async (id: string, title: string) => {
     try {
-      await showHUD("Restarted Habit: " + parseEmojiField(title).textWithoutEmoji);
+      await showHUD("Restarted Habit: " + stripPlannerEmojis(title).textWithoutEmoji);
       const [habit, error] = await axiosPromiseData(fetcher(`/planner/restart/habit/${id}`, { method: "POST" }));
       if (!habit || error) throw error;
       return habit;
@@ -101,7 +104,7 @@ export const useEventActions = () => {
 
   const handleStopHabit = async (id: string, title: string) => {
     try {
-      await showHUD("Stopped Habit: " + parseEmojiField(title).textWithoutEmoji);
+      await showHUD("Stopped Habit: " + stripPlannerEmojis(title).textWithoutEmoji);
       const [habit, error] = await axiosPromiseData(fetcher(`/planner/stop/habit/${id}`, { method: "POST" }));
       if (!habit || error) throw error;
 
@@ -114,7 +117,7 @@ export const useEventActions = () => {
 
   const handleStartOrRestartSmartHabit = async (lineageId: string, title: string) => {
     try {
-      await showHUD("Started Habit: " + parseEmojiField(title).textWithoutEmoji);
+      await showHUD("Started Habit: " + stripPlannerEmojis(title).textWithoutEmoji);
       const [habit, error] = await axiosPromiseData(
         fetcher(`/smart-habits/planner/${lineageId}/start`, { method: "POST" })
       );
@@ -129,7 +132,7 @@ export const useEventActions = () => {
 
   const handleStopSmartHabit = async (lineageId: string, title: string) => {
     try {
-      await showHUD("Stopped Habit: " + parseEmojiField(title).textWithoutEmoji);
+      await showHUD("Stopped Habit: " + stripPlannerEmojis(title).textWithoutEmoji);
       const [habit, error] = await axiosPromiseData(
         fetcher(`/smart-habits/planner/${lineageId}/stop`, { method: "POST" })
       );
@@ -150,7 +153,9 @@ export const useEventActions = () => {
 
     const hasRescheduleUnstarted = currentUser?.features.assistSettings.rescheduleUnstarted;
     const isEventManuallyStarted = event.assist?.manuallyStarted;
+
     const showStart = !isActive || (!!isActive && !!hasRescheduleUnstarted && !isEventManuallyStarted);
+
     const showRestartStop =
       !!isActive && (!hasRescheduleUnstarted || (!!hasRescheduleUnstarted && !!isEventManuallyStarted));